drf/requirements.txt
django~=3.0.0
djangorestframework~=3.11.0
httpie~=3.2.1